+++ lib/src/services_container.dart
@@ -1,0 +1,4 @@
+abstract interface class ServicesContainer {
+  Future<void> init();
+  Future<void> dispose();
+}